# [profiler] _RecordFunctionFast - faster python bindings for record_function