--- a/test/profiler/test_profiler.py
+++ b/test/profiler/test_profiler.py
@@ -1688,6 +1688,50 @@ assert KinetoStepTracker.current_step() == initial_step + 2 * niters
         finally:
             torch._C._profiler._set_record_concrete_inputs_enabled_val(True)
 
+    def test_record_function_fast(self):
+        x, y = (torch.rand((4, 4)) for _ in range(2))
+        with profile() as p:
+            for _ in range(4):
+                with torch._C._profiler_manual._RecordFunctionFast("add_test_fast_rf1"):
+                    x.add(y)
+
+        self.assertGreaterEqual(len([e for e in p.events() if e.name == "add_test_fast_rf1"]), 4)
+
+        with profile() as p:
+            cm = torch._C._profiler_manual._RecordFunctionFast("add_test_fast_rf2")
+            for _ in range(4):
+                with cm:
+                    x.add(y)
+
+        self.assertGreaterEqual(len([e for e in p.events() if e.name == "add_test_fast_rf2"]), 4)
+
+
+        with profile() as p:
+            cm = torch._C._profiler_manual._RecordFunctionFast("add_test_fast_rf3")
+            for _ in range(4):
+                try:
+                    with cm:
+                        x.add(y)
+                        raise ValueError()
+                        x.relu()
+                except ValueError:
+                    pass
+
+        self.assertGreaterEqual(len([e for e in p.events() if e.name == "add_test_fast_rf3"]), 4)
+        self.assertFalse(any((e.name and "relu" in e.name) for e in p.events()))
+
+        with profile() as p:
+            cm = torch._C._profiler_manual._RecordFunctionFast(None)
+            for i in range(4):
+                cm.set_name(f"add_test_fast_rf_{i}")
+                with cm:
+                    x.add(y)
+
+        self.assertTrue(any(e.name == "add_test_fast_rf_0" for e in p.events()))
+        self.assertTrue(any(e.name == "add_test_fast_rf_1" for e in p.events()))
+        self.assertTrue(any(e.name == "add_test_fast_rf_2" for e in p.events()))
+        self.assertTrue(any(e.name == "add_test_fast_rf_3" for e in p.events()))
+
 
 def find_node_with_name(nodes, name):
     for node in _utils.traverse_dfs(nodes):