… torch/_C/_profiler_manual.pyi