@davidberard98 2023-08-21
I moved the bindings from `torch._C._profiler_manual._RecordFunctionFast` to `torch._C._profiler._RecordFunctionFast`, following Ed's example from CapturedTraceback